[Rename] AgeGroup 파일 분리 (#35)
