+++ GetARock/GetARock/Network/Model/Band.swift
@@ -21,25 +21,3 @@ struct Band {
     /// 밴드의 자유소개 글입니다.
     var introduction: String
 }
-
-
-enum AgeGroup {
-    case lessThanTwenty
-    case twenties
-    case thirties
-    case forties
-    case fifties
-    case overThanSixty
-    
-    func toKorean() -> String {
-        switch self {
-        case .lessThanTwenty: return "20대 미만"
-        case .twenties: return "20대"
-        case .thirties: return "30대"
-        case .forties: return "40대"
-        case .fifties: return "50대"
-        case .overThanSixty: return "60대 이상"
-        }
-    }
-    
-}
